fixed code layouts

+++ math/miller_rabin.cpp
@@ -1,12 +1,12 @@
-int fastpow(int base, int d, int n) {
+int fastpow(int base, int d, int n){
     int ret = 1;
-    for (long long pow = base; d > 0; d >>= 1, pow = (pow * pow) % n)
+    for (int64 pow = base; d > 0; d >>= 1, pow = (pow * pow) % n)
         if (d & 1)
             ret = (ret * pow) % n;
     return ret;
 }
 
-bool miller_rabin(int n, int base) {
+bool miller_rabin(int n, int base){
     if (n <= 1) return false;
     if (n % 2 == 0) return n == 2;
 
@@ -17,16 +17,16 @@ bool miller_rabin(int n, int base) {
     if (base_d == 1) return true;
     int base_2r = base_d;
 
-    for (int i = 0; i < s; ++i) {
+    for (int i = 0; i < s; ++i){
         if (base_2r == 1) return false;
         if (base_2r == n - 1) return true;
-        base_2r = (long long)base_2r * base_2r % n;
+        base_2r = (int64)base_2r * base_2r % n;
     }
 
     return false;
 }
 
-bool isprime(int n) {
+bool isprime(int n){
     if (n == 2 || n == 7 || n == 61) return true;
     return miller_rabin(n, 2) && miller_rabin(n, 7) && miller_rabin(n, 61);
 }
